requirements-webapp.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv==1.0.1
psycopg2-binary==2.9.10
